fix draw cases and use seprate files for classes
