+++ lib/player.rb
@@ -1,0 +1,14 @@
+class Player
+  attr_reader :val, :name
+
+  def initialize(name, val)
+    @name = name
+    @val = val
+  end
+
+  def validate
+    return true if @name =~ /\w/
+
+    false
+  end
+end
